fix: GCC build cleanliness (queries.h array-bounds, data-types overflow)
Guard the QueryParams small-vector copy with a GCC-only diagnostic pragma (spurious GCC-14 -O2 -Warray-bounds false positive; clang/MSVC clean). Make the high-bit smallint test value an explicit static_cast to express the two's-complement intent (-Woverflow).

+++ tests/test-data-types.cpp
@@ -1009,7 +1009,7 @@ TEST_F(PostgreSQLDataTypesTest, TimestampTextFormatDeserialization) {
  */
 TEST_F(PostgreSQLDataTypesTest, HighBitValues) {
     // Create values with high bit set
-    smallint high_bit_smallint = 0x8000;               // -32768 in two's complement
+    smallint high_bit_smallint = static_cast<smallint>(0x8000); // -32768 in two's complement
     integer  high_bit_integer  = 0x80000000;           // -2147483648 in two's complement
     bigint   high_bit_bigint   = 0x8000000000000000LL; // Minimum for bigint
